接通databinding + room + liveDate Fixes: NMK- 描述: 无 测试步骤:

+++ app/src/main/java/com/simpure/expires/ui/PlaceAdapter.kt
@@ -86,7 +86,7 @@ class PlaceAdapter(
     }
 
     override fun getItemId(position: Int): Long {
-        return mCommodityList!![position].id
+        return mCommodityList!![position].id.toLong()
     }
 
     class ViewHolder(val binding: ItemCommodityOverviewBinding) :
